Recreated template helper recall scenario
This update recreates the scenario where the template helpers get
called again despite only a single field of a parent collection being
updated. In this case the project template which contains the task
template, is updated. But it is only the project name that is updated.
Despite this, the helpers and their relevant functions inside of the
tasks template are called again even though THE DOM IS NOT UPDATED
WITH IT