Restructured how data is retreived
This commit does two things. It changes the way data is retrived by
not using a custom ProjectModel object to which data is mapped but
instead using the accepted Meteor method which is to use _.extend
combined with the transform function in Meteor.Collections.

The second thing it does is keep the entire group of projects and
tasks under a larger with binding where the data at the top is a
reactive data source with data obtained from the DB.